hugolib: Pull all alias handling into one file
